+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Added
+- Added `UserInterfaceStyleOverridingElement` which allows child elements to have their `userInterfaceStyle` to be forced to light/dark. Additionally added a `overrideUserInterfaceStyle` convenience to `Element`.
 
 ### Removed
 - `AccessibilityElement.deprecated_accessibility(…)`. This was deprecated in September 2021, and renamed from .accessibility(…) to .deprecated_accessibility(…) in Oct 2024.